src/app/core/not-found/not-found.component.ts
import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
import { Title } from '@angular/platform-browser';

@Component({
  templateUrl: './not-found.component.html',
  styleUrls: ['./not-found.component.scss'],
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class NotFoundComponent implements OnInit {
  readonly pageTitle = 'Page Not Found';
  readonly erroMessage =
    'Oops! Something went wrong. Please check the URL and try again.';

  constructor(private readonly title: Title) {}

  ngOnInit(): void {
    this.title.setTitle(this.pageTitle);
  }
}
